requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.9
discord.py @ git+https://github.com/Rapptz/discord.py@45d498c1b76deaf3b394d17ccf56112fa691d160
dislash.py==1.4.9
idna==3.3
multidict==5.2.0
Pillow==9.0.0
pycparser==2.21
PyNaCl==1.4.0
pyocr==0.8.1
requests==2.26.0
six==1.16.0
tesseract==0.1.3
typing_extensions==4.0.1
urllib3==1.26.7
yarl==1.7.2
